+++ mutpy/operators/__init__.py
@@ -1,0 +1,46 @@
+from .arithmetic import *
+from .base import *
+from .decorator import *
+from .exception import *
+from .inheritance import *
+from .logical import *
+from .loop import *
+from .misc import *
+
+SuperCallingInsert = utils.get_by_python_version([
+    SuperCallingInsertPython27,
+    SuperCallingInsertPython35,
+])
+
+standard_operators = {
+    ArithmeticOperatorDeletion,
+    ArithmeticOperatorReplacement,
+    AssignmentOperatorReplacement,
+    BreakContinueReplacement,
+    ConditionalOperatorDeletion,
+    ConditionalOperatorInsertion,
+    ConstantReplacement,
+    DecoratorDeletion,
+    ExceptionHandlerDeletion,
+    ExceptionSwallowing,
+    HidingVariableDeletion,
+    LogicalConnectorReplacement,
+    LogicalOperatorDeletion,
+    LogicalOperatorReplacement,
+    OverriddenMethodCallingPositionChange,
+    OverridingMethodDeletion,
+    RelationalOperatorReplacement,
+    SliceIndexRemove,
+    SuperCallingDeletion,
+    SuperCallingInsert,
+}
+
+experimental_operators = {
+    ClassmethodDecoratorInsertion,
+    OneIterationLoop,
+    ReverseIterationLoop,
+    SelfVariableDeletion,
+    StatementDeletion,
+    StaticmethodDecoratorInsertion,
+    ZeroIterationLoop,
+}
