+++ mutpy/operators/loop.py
@@ -1,0 +1,45 @@
+import ast
+
+from mutpy.operators import copy_node, MutationOperator
+
+
+class OneIterationLoop(MutationOperator):
+    def one_iteration(self, node):
+        node.body.append(ast.Break())
+        return node
+
+    @copy_node
+    def mutate_For(self, node):
+        return self.one_iteration(node)
+
+    @copy_node
+    def mutate_While(self, node):
+        return self.one_iteration(node)
+
+
+class ReverseIterationLoop(MutationOperator):
+    @copy_node
+    def mutate_For(self, node):
+        old_iter = node.iter
+        node.iter = ast.Call(
+            func=ast.Name(id=reversed.__name__, ctx=ast.Load()),
+            args=[old_iter],
+            keywords=[],
+            starargs=None,
+            kwargs=None,
+        )
+        return node
+
+
+class ZeroIterationLoop(MutationOperator):
+    def zero_iteration(self, node):
+        node.body = [ast.Break()]
+        return node
+
+    @copy_node
+    def mutate_For(self, node):
+        return self.zero_iteration(node)
+
+    @copy_node
+    def mutate_While(self, node):
+        return self.zero_iteration(node)
