+++ tox.ini
@@ -5,7 +5,10 @@ envlist =
     coverage-report
 [testenv]
 deps=
-    -rrequirements/development.txt
+    tox
+    coverage
+extras=
+    pytest
 commands =
     coverage-erase: coverage erase
     test: coverage run --source=mutpy -m unittest discover -s mutpy/test
